+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/favorite/FavoriteRecordsSerializerTest.kt
@@ -2,10 +2,10 @@ package com.mapbox.search.favorite
 
 import com.mapbox.geojson.Point
 import com.mapbox.search.BaseTest
-import com.mapbox.search.ImageInfo
 import com.mapbox.search.SearchResultMetadata
 import com.mapbox.search.TestData
 import com.mapbox.search.common.RoutablePoint
+import com.mapbox.search.common.metadata.ImageInfo
 import com.mapbox.search.common.metadata.OpenHours
 import com.mapbox.search.common.metadata.OpenPeriod
 import com.mapbox.search.common.metadata.ParkingData