+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 - [CORE] Now `SearchSuggestion` and `SearchResult` types provide a new field `fullAddress` which contains formatted text address from backend.
 - [DISCOVER] Now `Discover` SDK provides callback-based alternatives for Kotlin suspend functions.
 - [DISCOVER] Now `DiscoverResult` provides a new field `formattedAddress`.
+- [PLACE AUTOCOMPLETE] Place Autocomplete SDK is available now. See docs for more information.
 
 ### Breaking changes
 - [CORE] `com.mapbox.search.CompletionCallback` has been moved to `com.mapbox.search.common.CompletionCallback`.
+- [CORE] `com.mapbox.search.ImageInfo` has been moved to `com.mapbox.search.common.metadata.ImageInfo`.
 
 ### Bug fixes
 - [CORE] Fixed `SearchSuggestion.name` and `SearchResult.name` formatting for address types.